refactor: move code for update-passwor users to auth
