+++ src/modules/auth/auth.controller.ts
@@ -1,4 +1,19 @@
-import { Body, Controller, Get, Post, Res, UseGuards } from '@nestjs/common';
+import { UpdatePasswordDto } from '@modules/users/dto/update-password.dto';
+import {
+  Body,
+  Controller,
+  Get,
+  HttpException,
+  HttpStatus,
+  Param,
+  ParseIntPipe,
+  Patch,
+  Post,
+  Req,
+  Res,
+  UseGuards,
+} from '@nestjs/common';
+import { Role } from '@src/common/enum/role.enum';
 import { Response } from 'express';
 import { AuthService } from './auth.service';
 import { ForgotPasswordDto } from './dto/forgotPassword.dto';
@@ -53,5 +68,24 @@ export class AuthController {
   @Post('reset-password')
   async resetPassword(@Body() resetPasswordDto: ResetPasswordDto) {
     return this.authService.resetPassword(resetPasswordDto);
+  }
+
+  @UseGuards(JwtAuthGuard)
+  @Patch(':id/password')
+  updatePassword(
+    @Body() updatePasswordDto: UpdatePasswordDto,
+    @Param('id', ParseIntPipe) id: number,
+    @Req() req,
+  ) {
+    const userId: number = req.user.id;
+    const isAdmin: boolean = req.user.role === Role.Admin;
+    const isUser: boolean = req.user.role === Role.User;
+
+    if (isAdmin || (isUser && userId === id))
+      return this.authService.updatePassword(id, updatePasswordDto);
+    throw new HttpException(
+      'Quyền truy cập bị hạn chế',
+      HttpStatus.BAD_REQUEST,
+    );
   }
 }
